requirements.txt
fastapi
uvicorn
pydantic
fastapi
uvicorn
pydantic
opencv-python
pillow
reportlab
torch
transformers
accelerate
